added support for more cmake files, fixed some type errors, cmake will be broken until order book is written

+++ utils/thread_utils_pin_cores.h
@@ -33,13 +33,13 @@ inline int pthread_setaffinity_np(
     cpu_set_t *cpu_set
 ) {
   thread_port_t mach_thread;
-  int core = 0;
+  size_t core = 0;
 
   for (core = 0; core < 8 * cpu_size; core++) {
     if (CPU_ISSET(core, cpu_set)) break;
   }
-  printf("binding to core %d\n", core);
-  thread_affinity_policy_data_t policy = { core };
+  printf("binding to core %d\n", (int)core);
+  thread_affinity_policy_data_t policy = { (int)core };
   mach_thread = pthread_mach_thread_np(thread);
   thread_policy_set(mach_thread, THREAD_AFFINITY_POLICY,
                     (thread_policy_t)&policy, 1);
